fix: FOUC (#8)
* refactor: replace ViewTransitions with ClientRouter in Layout.astro

This commit updates the Layout.astro file by replacing the ViewTransitions component with ClientRouter, enhancing the routing capabilities of the layout. This change aims to improve the overall navigation experience within the application.

* refactor: update theme color meta tags in Layout.astro

This commit comments out the previous theme color meta tags for light and dark modes and replaces them with a single empty theme color meta tag. This change simplifies the theme color configuration in the Layout.astro file.

* refactor: remove unused slot and clean up meta tags in Layout.astro

This commit removes the unused <slot name="head" /> from the Layout.astro file and further simplifies the theme color configuration by retaining only the empty theme color meta tag. These changes enhance the clarity and maintainability of the layout component.

* refactor: remove pagination links from Posts.astro layout

This commit eliminates the unused pagination links from the Posts.astro layout, simplifying the component and enhancing its maintainability. The changes contribute to a cleaner structure by focusing on essential elements within the layout.

* refactor: enhance Layout.astro with critical CSS and improve head slot usage

This commit adds inline minimal critical CSS to the Layout.astro file to mitigate flash of unstyled content (FOUC) during ViewTransitions. Additionally, it improves the usage of the head slot by ensuring it is included in the layout, enhancing the overall structure and maintainability of the component. The PostDetails and Posts layouts have also been updated for better meta tag formatting and pagination link management.

* Revert "refactor: enhance Layout.astro with critical CSS and improve head slot usage"

This reverts commit a4fb00e244833b09b85cdea1f39d015a7a917ac9.