refactor: update HTTP protocol titles and tags in blog post
This commit modifies the titles and tags in the blog post about HTTP protocols, changing them to a more consistent format. The titles for HTTP/1.0, HTTP/2.0, and HTTP/3.0 have been updated to HTTP/1, HTTP/2, and HTTP/3 respectively, and the corresponding tags have been adjusted to match this new format. These changes enhance clarity and maintainability of the content.